# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/seed": {
      "post": {
        "description": "Replaces the current in-memory vehicle catalog with the predefined dataset used for practice and onboarding. Requires an authenticated session with role `ADMIN`.",
        "operationId": "SeedController_populateDB",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Catalog dataset loaded successfully"
          },
          "401": {
            "description": "Missing or invalid session cookie"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Unexpected error while loading the catalog dataset"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Load the predefined catalog dataset",
        "tags": [
          "Catalog Administration"
        ]
      }
    },
...
  },
  "info": {
    "title": "NTT Demo Vehicle Catalog API",
    "description": "Backend API for the NTT Angular practice: authentication, vehicle catalog, brands, models, pagination and role-based access.\n\nRecommended flow in Swagger: call `POST /auth/login`, then use `GET /auth/me` and the protected endpoints normally. When the short-lived access cookie expires, call `POST /auth/refresh` and retry the failed request. Use `POST /auth/logout` to close the session cleanly.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
...
  }
}